+++ src/main/java/ru/autosome/commons/backgroundModel/mono/Background.java
@@ -65,7 +65,7 @@ public class Background implements BackgroundModel {
   }
 
   @Override
-  public double volume() {
+  public int volume() {
     return 1;
   }
